auth controller tests
